Move App to app.rs
