+++ cargo-lens/src/app.rs
@@ -1,0 +1,40 @@
+use std::marker::PhantomData;
+
+use ratatui::{
+    backend::Backend,
+    layout::{Constraint, Direction, Layout},
+    widgets::{Block, Borders, List, Paragraph},
+    Frame,
+};
+
+use crate::review_req_checklist::ReviewReqChecklist;
+
+pub struct App<B> {
+    pub list: ReviewReqChecklist,
+    _phantom: PhantomData<B>,
+}
+impl<B: Backend> App<B> {
+    pub fn new(list: ReviewReqChecklist) -> Self {
+        Self {
+            list,
+            _phantom: PhantomData,
+        }
+    }
+
+    pub fn render(&self, f: &mut Frame<B>) {
+        let chunks = Layout::default()
+            .direction(Direction::Horizontal)
+            .constraints([Constraint::Percentage(30), Constraint::Min(0)].as_ref())
+            .split(f.size());
+        let items = self.list.lines();
+
+        let block = Block::default().title("Checklist").borders(Borders::ALL);
+        let checklist = List::new(items).block(block);
+        f.render_widget(checklist, chunks[0]);
+        let block = Block::default().title("Info").borders(Borders::ALL);
+        let info =
+            Paragraph::new::<&str>(self.list.items.get(self.list.index).unwrap().info.as_ref())
+                .block(block);
+        f.render_widget(info, chunks[1]);
+    }
+}
